+++ config/application.rb
@@ -66,6 +66,7 @@ module DMPRoadmap
     # Requires server configuration and omniauth shibboleth provider configuration
     # See config/initializers/devise.rb
     config.shibboleth_enabled = false
+    config.openid_connect_enabled = true
 
     # Relative path to Shibboleth SSO Logout
     config.shibboleth_login = '/Shibboleth.sso/Login'